requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pgvector==0.3.5
pydantic==2.9.0
pydantic-settings==2.5.0
pyyaml==6.0.2
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
google-generativeai==0.8.0
python-multipart==0.0.9
aiofiles==24.1.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
